Finished sroll feature
itemPickerPromptStyle doesnt work due to it not being compatible with IOS. i had to implement 'justisifycontent 'center'' to make sure that it goes in the middle of the element. then i had to increase the height of the element above to be able to have the user input. although the scroll feature is still outside the box the function wokrs.